Quick fix for win32 systems
- Replaced ugly `Union` with `Any`
- Replaced all `Int64` with `Int`

+++ examples/sim1_complexity.jl
@@ -47,7 +47,7 @@ for is = 1:simlen, it = 1:trials
         M = sqrtM^2
         centers = square_grid(sqrtM, dist)
         φs = rand(M)
-        ids = ones(Int64,M)
+        ids = ones(Int, M)
         opt = FMMoptions(true, acc = Int(-log10(tol)), nx = div(sqrtM,2), method="pre")
     end
     tic()
